RapidPy/vrm_logger/requirements.txt
PySide6>=6.8,<6.9
pyserial>=3.5,<4
pyqtgraph>=0.13.7,<0.14
pyinstaller>=6.0,<7
